Compile fixes for builds on 64bit machines.
This mostly effects passing of plain types (integer) to
specialized parameters or return types from generic
functions (wrapping plain into ATOM types).

Use predefined type tag ids from known base types like lang|Int32
or lang|Bool.

+++ runtime/rt.h
@@ -17,9 +17,9 @@
 typedef struct ATOM ATOM;
 struct ATOM
 {
-  int typeid;
+  long typeid;
   union {
-    int   v_int;
+    long  v_int;
     float v_float;
     void* v_obj;                /* an heap allocated object */
   } u;
